pyproject.toml
[project]
name = "imalink"
version = "0.1.0"
description = "ImaLink - Image gallery and management system"
readme = "README.md"
requires-python = ">=3.13.7"
dependencies = [
    "aiofiles>=24.1.0",
    "alembic>=1.16.5",
    "cairosvg>=2.8.2",
    "email-validator>=2.3.0",
    "fastapi>=0.118.0",
    "flet>=0.28.3",
    "imagehash>=4.3.2",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.3.0",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.37.0",
    "psycopg2-binary>=2.9.9",
]

[dependency-groups]
dev = [
    "black==23.10.1",
    "httpx==0.25.2",
    "isort==5.12.0",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "requests>=2.32.5",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
